feat: Implement Next.js Image for avatar, refactor mock data location, update page metadata, and remove unnecessary 'use client' directives. (#57)
