+++ src/app/layout.tsx
@@ -19,8 +19,13 @@ const poppins = Poppins({
 });
 
 export const metadata: Metadata = {
-  title: 'Millinks',
-  description: 'Generated by create next app',
+  title: 'Robson Natanael | Millinks',
+  description: 'Links e contatos de Robson Natanael',
+  openGraph: {
+    title: 'Robson Natanael | Millinks',
+    description: 'Links e contatos de Robson Natanael',
+    type: 'website',
+  },
 };
 
 export default function RootLayout({
