+++ src/app/page.tsx
@@ -1,8 +1,6 @@
-'use client';
-
 import Layout from '@/components/Layout';
 import Header from '@/container/Home/Header';
-import { data } from './api/mock';
+import { data } from '@/data/mock';
 import Nav from '@/container/Home/Nav';
 import Footer from '@/components/Footer';
