+++ src/components/Footer/index.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import Image from 'next/image';
 
 import { Box, Container } from '@mui/material';
