+++ src/container/Home/Nav/index.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { Button, Container } from '@mui/material';
 
 import { LinksProps } from './props';
